Embed modules within chunks (#11)
There are cases where you can have modules with the same `readableIdentifier` but different content.  (I don't think this will happen with source modules, but I've observed with with webpack runtime modules.)  This change embeds the modules within the chunks so that it is clear which instance of each module is associated with which chunk.